float -> double since reqs require double-precision

+++ main.c
@@ -4,7 +4,7 @@ int main() { // take in inputs here
     int rows, columns; // size of matrix
     scanf("%d %d", &rows, &columns); // user input (first line)
 
-    float matrix[rows][columns]; // store the image pixel data
+    double matrix[rows][columns]; // store the image pixel data
 
     return 0;
 }
